docker/kafka_apps/2_consumers/requirements.txt
kafka-python
confluent-kafka
faker
fastavro
requests